requirements.txt
beautifulsoup4==4.6.0
nose==1.3.7
requests==2.18.4
cryptography==2.2.2
cffi==1.11.5
asn1crypto==0.24.0
pyOpenSSL==17.5.0
pycparser==2.18
PySocks==1.6.7
stem==1.6.0
Django==2.0.1
psycopg2==2.7.3.2
django-model-utils==3.1.1
selenium==3.9.0
celery==4.1.0
django-celery-beat==1.1.1
redis==2.10.6
dateparser==0.7.0
ipython==6.2.1
ipdb==0.10.3
gunicorn == 19.7.1
django-autocomplete-light == 3.3.2